shared-music-play/pyproject.toml
[tool.poetry]
name = "shared-music-play"
version = "0.1.0"
description = ""
authors = ["Nishanbaev Timur <redbuzzzzzz@gmail.com>"]
readme = "README.md"
packages = [{include = "shared_music_play"}]

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.5.2"
sqlparse = "0.4.3"
tzdata = "2022.7"
django = "4.1.4"
psycopg2-binary = "2.9.5"
pre-commit = "2.20.0"
black = "22.12.0"
djangorestframework = "3.14.0"
requests = "^2.28.2"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
drf-nested-routers = "^0.93.4"
pyjwt = "^2.6.0"
django-cors-headers = "^3.14.0"
google-api-python-client = "^2.85.0"
uwsgi = "^2.0.21"
awscli = "^1.27.133"
daphne = "^4.0.0"
channels = "^4.0.0"
channels-postgres = "^0.0.5"
django-environ = "^0.10.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
